bullets flies almost correctly :) added set towers and monsters data

+++ src/Monster.js
@@ -1,19 +1,35 @@
 import Phaser from 'phaser';
 
 export default class Monster extends Phaser.Physics.Arcade.Image {
-  constructor(scene, x, y) {
+  constructor(scene, x, y, data) {
     super(scene, x, y, 'monster');
 
     this.scene = scene;
     this.x = x;
     this.y = y;
-    this.hp = 6;
+    this.name;
+    this.hp;
+    this.speed;
+    this.armor;
+    this.magicResistance;
+    this.type;
+    this.ability;
 
     this.selected = false;
   }
 
   setSelected(selected) {
     this.selected = selected;
+  }
+
+  setParams(data) {
+    this.name = data.name;
+    this.hp = data.hp;
+    this.speed = data.speed;
+    this.armor = data.armor;
+    this.magicResistance = data.magicResistance;
+    this.type = data.type;
+    this.ability = data.ability;
   }
 
   update() {
@@ -25,6 +41,10 @@ export default class Monster extends Phaser.Physics.Arcade.Image {
         this.scene.registry.get('points')[6].y *
           this.scene.registry.get('frameSize')
     ) {
+      this.delete();
+    }
+
+    if (this.hp <= 0) {
       this.delete();
     }
   }
